fix: fixing lytics user_id and anonymousId mapping (#3745)
* fix: fixing lytics user_id and anonymousId mapping

* fix: adding mapping config for identify call and clean up of native code

* fix: adding user and anonymousId mapping for other calls as well

---------
