+++ src/cdk/v2/destinations/lytics/procWorkflow.yaml
@@ -5,9 +5,12 @@ bindings:
     path: ../../../../v0/util
   - name: removeUndefinedAndNullValues
     path: ../../../../v0/util
+  - name: constructPayload
+    path: ../../../../v0/util
   - path: ../../bindings/jsontemplate
   - name: defaultRequestConfig
     path: ../../../../v0/util
+  - path: ./config
 
 steps:
   - name: validateInput
@@ -24,20 +27,19 @@ steps:
     condition: $.context.messageType === {{$.EventType.IDENTIFY}}
     template: |
       const flattenTraits = $.flattenJson(.message.traits ?? .message.context.traits);
-      $.context.payload = .message.({
+      const payload = $.constructPayload(.message, $.CUSTOMER_PROPERTIES_CONFIG);
+      $.context.payload = {
         ...flattenTraits,
-        first_name: {{{{$.getGenericPaths("firstName")}}}},
-        last_name: {{{{$.getGenericPaths("lastName")}}}},
-        user_id: {{{{$.getGenericPaths("userId")}}}}
-      })
+        ...payload,
+      }
     else:
       name: payloadForOthers
       template: |
         const flattenProperties = $.flattenJson(.message.properties);
+        const customerPropertiesInfo = $.constructPayload(.message, $.CUSTOMER_PROPERTIES_CONFIG);
         $.context.payload = .message.({
           ...flattenProperties,
-          first_name: .properties.firstName ?? .properties.firstname,
-          last_name: .properties.lastName ?? .properties.lastname
+          ...customerPropertiesInfo
         })
   - name: trackPayload
     condition: $.context.messageType === {{$.EventType.TRACK}}
