--- src/v0/destinations/lytics/data/LYTICSIdentifyConfig.json
@@ -1,19 +1,0 @@
-[
-  {
-    "destKey": "user_id",
-    "sourceKeys": [
-      "userId",
-      "traits.userId",
-      "traits.id",
-      "context.traits.userId",
-      "context.traits.id",
-      "anonymousId"
-    ],
-    "required": false
-  },
-  {
-    "destKey": "",
-    "sourceKeys": ["traits", "context.traits"],
-    "required": false
-  }
-]
