--- src/v0/destinations/lytics/data/LYTICSPageScreenConfig.json
@@ -1,12 +1,0 @@
-[
-  {
-    "destKey": "event",
-    "sourceKeys": "name",
-    "required": false
-  },
-  {
-    "destKey": "",
-    "sourceKeys": "properties",
-    "required": false
-  }
-]
