--- src/v0/destinations/lytics/data/LYTICSTrackConfig.json
@@ -1,12 +1,0 @@
-[
-  {
-    "destKey": "_e",
-    "sourceKeys": "event",
-    "required": false
-  },
-  {
-    "destKey": "",
-    "sourceKeys": "properties",
-    "required": false
-  }
-]
